make dependensy of dets file creation not critical

+++ lib/unleash/cache.ex
@@ -4,7 +4,10 @@ defmodule Unleash.Cache do
   threads to read the feature flag values concurrently on top of minimizing
   network calls
   """
+
   alias Unleash.Config
+
+  require Logger
 
   @cache_table_name :unleash_cache
   @dets_table_name :unleash_dets_cache
@@ -16,9 +19,14 @@ defmodule Unleash.Cache do
   """
   def init(existing_features \\ [], table_name \\ @cache_table_name) do
     :ets.new(table_name, [:named_table, read_concurrency: true])
-    {:ok, _pid} = :dets.open_file(@dets_table_name, file: Config.dets_file())
-
+    case :dets.open_file(@dets_table_name, file: Config.dets_file()) do
+      {:ok, _pid} ->
+        Logger.info("Opened DETS file #{Config.dets_file()} successfully")
+      {:error, reason} ->
+        Logger.error("Failed to open DETS file: #{inspect(reason)}")
+    end
     upsert_features(existing_features, table_name)
+
   end
 
   @doc """
@@ -66,7 +74,7 @@ defmodule Unleash.Cache do
   def upsert_features(features, table_name \\ @cache_table_name)
 
   def upsert_features([], table_name) do
-    if :dets.is_dets_file(Config.dets_file()) do
+    if dets_file_exists?() do
       :ets.delete_all_objects(table_name)
       :ets.from_dets(table_name, @dets_table_name)
     end
@@ -79,10 +87,6 @@ defmodule Unleash.Cache do
       upsert_feature(feature.name, feature, table_name)
     end)
 
-    :ok = :dets.delete_all_objects(@dets_table_name)
-    :ok = :dets.from_ets(@dets_table_name, table_name)
-    # Add this line to force sync to disk
-    :ok = :dets.sync(@dets_table_name)
   end
 
   defp upsert_feature(name, value, table_name) when is_binary(name) do
@@ -94,13 +98,20 @@ defmodule Unleash.Cache do
   end
 
   def dets_file_exists? do
-    :dets.is_dets_file(Config.dets_file())
+    :dets.is_dets_file(Config.dets_file()) == true
   end
 
   def read_features_from_dets(table_name \\ @cache_table_name) do
     :ets.delete_all_objects(table_name)
     :ets.from_dets(table_name, @dets_table_name)
     get_features(table_name)
+  end
+
+  def write_features_to_dets(table_name \\ @cache_table_name) do
+    :ok = :dets.delete_all_objects(@dets_table_name)
+    :ok = :dets.from_ets(@dets_table_name, table_name)
+    # Add this line to force sync to disk
+    :ok = :dets.sync(@dets_table_name)
   end
 
 end
